refactored the entire eleventy setup to be a lot cleaner and separated
