--- 11ty/lists.js
@@ -1,5 +1,0 @@
-function head(list, n) {
-  return list.slice(0, n);
-}
-
-module.exports = { head };
