--- 11ty/readTime.js
@@ -1,3 +1,0 @@
-module.exports = (str) => {
-  return `${Math.ceil(str.split(" ").length / 200)} min read`;
-};
