--- 11ty/readableDate.js
@@ -1,7 +1,0 @@
-module.exports = (date) => {
-  return new Date(date).toLocaleDateString("en-us", {
-    year: "numeric",
-    month: "long",
-    day: "numeric",
-  });
-};
