+++ config/libraries.js
@@ -2,10 +2,14 @@ const anchor = require("markdown-it-anchor");
 const shiki = require("markdown-it-shiki").default;
 const container = require("markdown-it-container");
 
-module.exports = (md) =>
-  md
+// Use shiki as the highlighter and aadd anchor tags
+function md(mdit) {
+  mdit
     .use(container, "info")
     .use(anchor, {
       permalink: anchor.permalink.ariaHidden({ symbol: "#" }),
     })
     .use(shiki, { theme: "dracula" });
+}
+
+module.exports = { md };
