+++ src/_includes/components/tag-list.njk
@@ -1,7 +1,5 @@
-{% set _tags = tags | head(5) %}
 <ul role="list" class="list-group list-group--scale | cluster items-center --gap-00">
-  {% for tag in _tags %}
+  {% for tag in tags %}
     <li class="tag | click-area | size-00" data-selected="{{ selected === tag }}"><a href="/tags/{{ tag }}/">{{ tag }}</a></li>
   {% endfor %}
-  {% if tags | length > 4 %}<li><a href="/tag-list" class="size-00 redirect">all topics</a></li>{% endif %}
 </ul>
