+++ src/_includes/layouts/post.njk
@@ -11,7 +11,7 @@
       {{ content | safe }}
     </article>
 
-    {% set toc = content | headers %}
+    {% set toc = content | getPostHeaders %}
     {% if toc | length %}
     <nav class="toc | sticky | stack --gap-000 self-start">
     <span class="size-00 uppercase sans-serif bold">Table of contents</span>
