+++ src/tag-list.njk
@@ -6,10 +6,10 @@ width: 3
 ---
 
 <ul>
-  {% for tag in collections.posts | getAllTagsWithCount %}
+  {% for tag, count in collections.posts | getAllTags(true) %}
   <li>
-    <a href="/tags/{{ tag[0] }}/">{{ tag [0]}}</a
-    ><span class="size-00"> ({{ tag[1] }} articles)</span>
+    <a href="/tags/{{ tag }}/">{{ tag }}</a>
+    <span class="size-00"> ({{ count }} articles)</span>
   </li>
   {% endfor %}
 </ul>
