+++ src/tags.njk
@@ -13,8 +13,11 @@ eleventyComputed:
   title: Tagged “{{ tag }}”
 ---
 
-{% set tags = collections.posts | getAllTags | head(5) %}
-{% set selected = tag %}
-{% include "components/tag-list.njk" %}
+{% set tags = collections.posts | getAllTags | head(6) %}
+{% set selected = tag[0] %}
+<div class="cluster items-center --gap-00">
+  {% include "components/tag-list.njk" %}
+  <a href="/tag-list" class="size-00 redirect">all topics</a>
+</div>
 {% set posts = collections[ tag ] or [] %}
 {% include "components/post-list.njk" %}
